+++ lib/gn_list_resolver/result_processor.rb
@@ -5,7 +5,9 @@ module GnListResolver
   class ResultProcessor
     attr_reader :input, :writer
 
-    def initialize(writer, stats)
+    def initialize(writer, stats, with_classification = false)
+      @with_classification = with_classification
+      @parser = ScientificNameParser.new
       @stats = stats
       @writer = writer
       @input = {}
@@ -20,26 +22,28 @@ module GnListResolver
 
     private
 
-    def rubyfy(result)
-      JSON.parse(result, symbolize_names: true)
-    end
-
     def write_empty_result(datum)
       @stats.stats[:matches][MATCH_TYPE_EMPTY] += 1
       @stats.stats[:resolved_records] += 1
-      res = @original_data[datum.supplied_id]
-      res += [MATCH_TYPES[MATCH_TYPE_EMPTY], datum.supplied_input, nil,
-              nil,
-              @input[datum.supplied_id][:rank],
-              nil,
-              nil, nil, nil]
+      res = compile_empty_result(datum)
       @writer.write(res)
+    end
+
+    def compile_empty_result(datum)
+      res = @original_data[datum.supplied_id]
+      res += [MATCH_TYPES[MATCH_TYPE_EMPTY], 0, datum.supplied_input,
+              nil, nil, nil, nil,
+              @input[datum.supplied_id][:rank],
+              nil, nil, nil, nil, nil]
+      res <<  nil if @with_classification
+      res
     end
 
     def write_result(datum)
       collect_stats(datum)
+      match_size = datum.results.size
       datum.results.each do |result|
-        @writer.write(compile_result(datum, result))
+        @writer.write(compile_result(datum, result, match_size))
       end
     end
 
@@ -54,28 +58,56 @@ module GnListResolver
       @stats.stats[:resolved_records] += 1
     end
 
-    def compile_result(datum, result)
-      @original_data[datum.supplied_id] + prepare_data(datum, result)
+    def compile_result(datum, result, match_size)
+      @original_data[datum.supplied_id] + prepare_data(datum,
+                                                       result, match_size)
     end
 
     # rubocop:disable Metrics/AbcSize
 
-    def prepare_data(datum, result)
-      [MATCH_TYPES[result.match_type.kind.to_sym],
-       datum.supplied_input, result.name.value,
-       result.canonical_name.value, @input[datum.supplied_id][:rank],
-       matched_rank(result),
-       result.synonym,
-       result.name.value, # TODO: should be `current_name_string` field
-       result.match_type.edit_distance,
-       result.score.value ? result.score.value.round(3) : nil,
-       result.taxon_id]
+    def prepare_data(datum, result, match_size)
+      res = [MATCH_TYPES[result.match_type.kind.to_sym], match_size,
+             datum.supplied_input, result.name.value,
+             canonical(datum.supplied_input), result.canonical_name.value,
+             result.match_type.edit_distance, @input[datum.supplied_id][:rank],
+             matched_rank(result), result.synonym, current_name(result),
+             result.score.value ? result.score.value.round(3) : nil,
+             result.taxon_id]
+      res << classification(result) if @with_classification
+      res
     end
 
     # rubocop:enable all
 
+    def current_name(result)
+      result.accepted_name.name.value || result.name.value
+    end
+
+    def canonical(name_string)
+      parsed = @parser.parse(name_string)[:scientificName]
+      return nil if parsed[:canonical].nil? || parsed[:hybrid]
+      parsed[:canonical]
+    rescue StandardError
+      @parser = ScientificNameParser.new
+      nil
+    end
+
     def matched_rank(result)
       result.classification.path_ranks.split("|").last
     end
+
+    # rubocop:disable Metrics/AbcSize
+
+    def classification(result)
+      return nil if result.classification.path.to_s.strip == ""
+      path = result.classification.path.split("|")
+      ranks = result.classification.path_ranks.split("|")
+      if path.size == ranks.size
+        path = path.zip(ranks).map { |e| "#{e[0]}(#{e[1]})" }
+      end
+      path.join(", ")
+    end
+
+    # rubocop:enable all
   end
 end