+++ spec/features/checklist_resolver_spec.rb
@@ -2,17 +2,20 @@
 
 describe "features" do
   context "resolving variety of csv files" do
-    %i[all_fields sciname sciname_auth sciname_rank csv_relaxed].each do |input|
+    %i[single_field empty_row all_fields sciname sciname_auth
+       sciname_rank csv_relaxed].each do |input|
       context input do
         it "resolves #{input}" do
           opts = { output: "/tmp/#{input}-processed.csv",
                    input: FILES[input],
                    data_source_id: 1,
-                   skip_original: true,
+                   with_classification: [true, false].sample,
+                   skip_original: [true, false].sample,
                    debug: true }
-          FileUtils.rm(opts[:output]) if File.exist?(opts[:output])
           GnListResolver.run(opts)
           expect(File.exist?(opts[:output])).to be true
+          expect(uniform_rows?(opts[:output])).to be true
+          FileUtils.rm(opts[:output])
         end
       end
     end
@@ -34,6 +37,21 @@ describe "features" do
     end
   end
 
+  context "matchSize" do
+    it "shows how many matches happen" do
+      opts = { output: "/tmp/output.csv",
+               input: FILES[:sciname],
+               data_source_id: 1, skip_original: true }
+      GnListResolver.run(opts)
+      expect(File.exist?(opts[:output])).to be true
+      res = CSV.open(opts[:output], col_sep: "\t", headers: true).map do |r|
+        r["matchSize"]
+      end.uniq.sort
+      expect(res).to eq %w[0 1 2 479]
+      FileUtils.rm(opts[:output])
+    end
+  end
+
   context "combining acceptedName output" do
     it "gives accepted name for all matches"
     # do
@@ -41,10 +59,12 @@ describe "features" do
     #            input: FILES[:sciname],
     #            data_source_id: 1, skip_original: true }
     #   GnListResolver.run(opts)
+    #   require "byebug"; byebug
     #   CSV.open(opts[:output], col_sep: "\t", headers: true).each do |r|
-    #     next unless r["matchedEditDistance"] == "0"
-    #     expect(r["matchedName"].size).to be > 1
-    #     expect(r["acceptedName"].size).to be > 1
+    #     next unless r["synonymStatus"] == "true"
+    #     expect(r["matchedName"].strip).to_not eq ""
+    #     expect(r["acceptedName"].strip).to_not eq ""
+    #     expect(r["matchedName"]).to_not eq r["acceptedName"]
     #   end
     #   FileUtils.rm(opts[:output])
     # end
@@ -76,6 +96,7 @@ describe "features" do
         expect(r["inputName"]).to eq "Animalia"
         expect(r["acceptedName"]).to eq "Animalia"
       end
+      FileUtils.rm(opts[:output])
     end
 
     it "breaks without alternative headers" do
@@ -83,6 +104,7 @@ describe "features" do
                input: FILES[:no_name],
                data_source_id: 1, skip_original: true }
       expect { GnListResolver.run(opts) }.to raise_error GnListResolverError
+      FileUtils.rm(opts[:output])
     end
 
     it "uses complex alternative headers" do
@@ -108,7 +130,25 @@ describe "features" do
                data_source_id: 1, skip_original: true }
       GnListResolver.run(opts) { "STOP" }
       lines_num = File.readlines(opts[:output]).size
-      expect(lines_num).to eq 1005
+      expect(lines_num).to eq 1004
+      FileUtils.rm(opts[:output])
+    end
+  end
+
+  context "with classification" do
+    it "creates classification if the option is true" do
+      opts = { output: "/tmp/output.csv",
+               input: FILES[:sciname],
+               data_source_id: 1, skip_original: true,
+               with_classification: true }
+      GnListResolver.run(opts)
+      classification_exists = false
+      CSV.open(opts[:output], col_sep: "\t", headers: true).each do |r|
+        if r["classification"] && r["classification"].split(", ").size > 1
+          classification_exists = true
+        end
+      end
+      expect(classification_exists).to be true
       FileUtils.rm(opts[:output])
     end
   end